requirements.txt
beautifulsoup4==4.13.4
faiss_cpu==1.7.4
flagembedding==1.3.4
fpdf==1.7.2
fpdf2==2.8.2
huggingface_hub==0.30.2
mistral_common==1.5.4
mistral_inference==1.6.0
msal==1.32.3
numpy==2.2.5
Office365_REST_Python_Client==2.6.1
pandas==2.2.3
pdfplumber==0.10.3
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.3.0
sentence_transformers==4.1.0
spacy==3.7.5
spacy_layout==0.0.12
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3